requirement.txt
Flask==2.1.2
Werkzeug==2.1.2
azure-storage-blob==12.13.1
mysql-connector-python==8.0.29
PyJWT==2.4.0
redis==4.3.4
